# docs/swagger.yaml
definitions:
  BadRequestAPIResponse:
    properties:
      data: {}
      meta:
        properties:
          message:
            example: Bad request
            type: string
          status:
            example: 400
            type: integer
        type: object
    type: object
  InternalServerErrorAPIResponse:
    properties:
      meta:
        properties:
          message:
            example: Internal server error
            type: string
          status:
            example: 500
            type: integer
        type: object
    type: object
  LoginRequest:
    properties:
      deviceinfo:
        example: postmanstudent123x0==
        type: string
      nisn:
        example: "2024000579"
        type: string
      password:
        example: password123
        type: string
    required:
    - nisn
    - password
    type: object
  LoginSuccessResponse:
    properties:
      data:
        properties:
          access_token:
            example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
            type: string
          created_at:
            example: "0001-01-01T00:00:00Z"
            type: string
          device_info:
            example: postmanstudent123x0==
            type: string
          last_activity:
            example: "0001-01-01T00:00:00Z"
            type: string
          name:
            example: Eko Saputra
            type: string
          nisn:
            example: "2024000579"
            type: string
          session_id:
            example: b5e882d0eb21c5ba7752d2b8f216ccad
            type: string
          user_id:
            example: 003061ee-97ff-4d00-8155-f4bf15e319dd
            type: string
        type: object
      meta:
        properties:
          message:
            example: login success
            type: string
          status:
            example: 200
            type: integer
        type: object
    type: object
  SuccessOnlyAPIResponse:
    properties:
      meta:
        properties:
          message:
            example: Operation successful
            type: string
          status:
            example: 200
            type: integer
        type: object
    type: object
  UnauthorizedAPIResponse:
    properties:
      meta:
        properties:
          message:
            example: Unauthorized access
            type: string
          status:
            example: 401
            type: integer
        type: object
    type: object
  config.HealthResponse:
    properties:
      database:
        example: connected
        type: string
      redis:
        example: connected
        type: string
      status:
        example: healthy
        type: string
      timestamp:
        example: "2024-01-01T12:00:00Z"
        type: string
    type: object
  config.RootResponse:
    properties:
      env:
        example: dev
        type: string
      message:
        example: Presensi SMK API is running!
        type: string
      status:
        example: success
        type: string
      version:
        example: 1.0.0
        type: string
    type: object
info:
  contact: {}
  description: Ini adalah API untuk ujikom mif 2025
  termsOfService: http://swagger.io/terms/
  title: UJIKOM API
  version: "1.0"
paths:
  /:
    get:
      consumes:
      - application/json
      description: Returns basic information about the API including status, version,
        and environment
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/config.RootResponse'
      summary: Get API Information
      tags:
      - System
  /student/login:
    post:
      consumes:
      - application/json
      description: Authenticate student dengan NISN dan password
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/LoginSuccessResponse'
        "400":
          description: Bad request - invalid input
          schema:
            $ref: '#/definitions/BadRequestAPIResponse'
        "401":
          description: Unauthorized - invalid credentials
          schema:
            $ref: '#/definitions/UnauthorizedAPIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/InternalServerErrorAPIResponse'
      security:
      - ApiKeyAuth: []
      summary: Student Login
      tags:
      - Auth
  /student/logout:
    post:
      consumes:
      - application/json
      description: Logout student dan hapus session yang aktif
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/SuccessOnlyAPIResponse'
        "401":
          description: Unauthorized - invalid or expired token
          schema:
            $ref: '#/definitions/UnauthorizedAPIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/InternalServerErrorAPIResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Student Logout
      tags:
      - Auth
  /ujikom/api/health:
    get:
      consumes:
      - application/json
      description: Returns the health status of the API including database and Redis
        connectivity
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/config.HealthResponse'
      summary: Health Check
      tags:
      - System
swagger: "2.0"